Move HTTP from fdbclient/ to fdbrpc/.
