+++ fdbclient/S3BlobStore.actor.cpp
@@ -20,8 +20,8 @@
 
 #include "fdbclient/S3BlobStore.h"
 
-#include "fdbclient/md5/md5.h"
-#include "fdbclient/libb64/encode.h"
+#include "fdbrpc/md5/md5.h"
+#include "fdbrpc/libb64/encode.h"
 #include "fdbclient/sha1/SHA1.h"
 #include <time.h>
 #include <iomanip>
