+++ fdbclient/include/fdbclient/ClientKnobs.h
@@ -217,12 +217,7 @@ public:
 	int64_t CSI_SIZE_LIMIT;
 	double CSI_STATUS_DELAY;
 
-	int HTTP_SEND_SIZE;
-	int HTTP_READ_SIZE;
-	int HTTP_VERBOSE_LEVEL;
-	std::string HTTP_REQUEST_ID_HEADER;
 	bool HTTP_REQUEST_AWS_V4_HEADER; // setting this knob to true will enable AWS V4 style header.
-	bool HTTP_RESPONSE_SKIP_VERIFY_CHECKSUM_FOR_PARTIAL_CONTENT; // skip verify md5 checksum for 206 response
 	std::string BLOBSTORE_ENCRYPTION_TYPE;
 	int BLOBSTORE_CONNECT_TRIES;
 	int BLOBSTORE_CONNECT_TIMEOUT;
