+++ fdbclient/include/fdbclient/S3BlobStore.h
@@ -26,7 +26,7 @@
 #include "flow/Net2Packet.h"
 #include "fdbclient/Knobs.h"
 #include "flow/IRateControl.h"
-#include "fdbclient/HTTP.h"
+#include "fdbrpc/HTTP.h"
 #include "fdbclient/JSONDoc.h"
 
 // Representation of all the things you need to connect to a blob store instance with some credentials.
