+++ fdbrpc/include/fdbrpc/libb64/cencode.h
