+++ fdbrpc/include/fdbrpc/libb64/decode.h
