+++ fdbrpc/include/fdbrpc/libb64/encode.h
