+++ fdbrpc/libb64/cencode.c
@@ -5,7 +5,7 @@ This is part of the libb64 project, and has been placed in the public domain.
 For details, see http://sourceforge.net/projects/libb64
 */
 
-#include "fdbclient/libb64/cencode.h"
+#include "fdbrpc/libb64/cencode.h"
 
 const int CHARS_PER_LINE = 72;
